refactor, indices folder
