+++ graph_query.h
@@ -28,8 +28,8 @@
 #include <boost/graph/vf2_sub_graph_iso.hpp>
 #include <boost/date_time/posix_time/posix_time.hpp>
 #include "get_traces_by_structure.h"
-#include "folders_index_query.h"
-#include "query_bloom_index.h"
+#include "indices/folders_index_query.h"
+#include "indices/query_bloom_index.h"
 #include "common.h"
