+++ indices/folders_index_query.cc
